further refactoring, nitro signator and verifier for payment state authentication
