--- services/payments/qldb_types.go
@@ -1,46 +1,0 @@
-package payments
-
-import (
-	"fmt"
-	"time"
-
-	qldbTypes "github.com/aws/aws-sdk-go-v2/service/qldb/types"
-	paymentLib "github.com/brave-intl/bat-go/libs/payments"
-)
-
-// QLDBPaymentTransitionHistoryEntryBlockAddress defines blockAddress data for
-// qldbPaymentTransitionHistoryEntry.
-type QLDBPaymentTransitionHistoryEntryBlockAddress struct {
-	StrandID   string `ion:"strandId"`
-	SequenceNo int64  `ion:"sequenceNo"`
-}
-
-// QLDBPaymentTransitionHistoryEntryHash defines hash for qldbPaymentTransitionHistoryEntry.
-type QLDBPaymentTransitionHistoryEntryHash []byte
-
-// qldbPaymentTransitionHistoryEntrySignature defines signature for qldbPaymentTransitionHistoryEntry.
-type qldbPaymentTransitionHistoryEntrySignature []byte
-
-// QLDBPaymentTransitionHistoryEntryMetadata defines metadata for qldbPaymentTransitionHistoryEntry
-type QLDBPaymentTransitionHistoryEntryMetadata struct {
-	ID      string    `ion:"id"`
-	TxID    string    `ion:"txId"`
-	TxTime  time.Time `ion:"txTime"`
-	Version int64     `ion:"version"`
-}
-
-// QLDBPaymentTransitionHistoryEntry defines top level entry for a QLDB transaction.
-type QLDBPaymentTransitionHistoryEntry struct {
-	BlockAddress QLDBPaymentTransitionHistoryEntryBlockAddress `ion:"blockAddress"`
-	Hash         QLDBPaymentTransitionHistoryEntryHash         `ion:"hash"`
-	Data         paymentLib.PaymentState                                  `ion:"data"`
-	Metadata     QLDBPaymentTransitionHistoryEntryMetadata     `ion:"metadata"`
-}
-
-// ValueHolder converts a QLDBPaymentTransitionHistoryEntry into a QLDB SDK ValueHolder.
-func (b *QLDBPaymentTransitionHistoryEntryBlockAddress) ValueHolder() *qldbTypes.ValueHolder {
-	stringValue := fmt.Sprintf("{strandId:\"%s\",sequenceNo:%d}", b.StrandID, b.SequenceNo)
-	return &qldbTypes.ValueHolder{
-		IonText: &stringValue,
-	}
-}
